@@ -3,9 +3,10 @@ package com.adxcorp.adxdev;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.widget.RelativeLayout;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.adxcorp.gdpr.ADXGDPR;
 import com.adxcorp.nativead.NativeAdFactory;
